S11P12B209-109 Refactor:AuthController 추가 및 수정, Security Config 수정, token 만료기간 수정

+++ backend/src/main/java/com/common/togather/db/repository/MemberRepository.java
@@ -7,9 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-//    Optional<Member> findByEmail(String email);
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 
     Member findByEmail(String email);
